Prepare CoreReplApp for partial extraction into standalone classes
Promote private nested types (ActiveRoutingGraph, PrefixTemplate,
ContextValidationOutcome, AmbientCommandOutcome) to standalone internal
files. Promote key methods from private to internal for cross-class
access. Merge OptionParsing partial (62 lines) into Execution partial.
