+++ src/Repl.Core/CoreReplApp.Execution.cs
@@ -427,7 +427,7 @@ public sealed partial class CoreReplApp
 		"Maintainability",
 		"MA0051:Method is too long",
 		Justification = "Execution path intentionally keeps validation, binding, middleware and rendering in one place.")]
-	private async ValueTask<(int ExitCode, bool EnterInteractive)> ExecuteMatchedCommandAsync(
+	internal async ValueTask<(int ExitCode, bool EnterInteractive)> ExecuteMatchedCommandAsync(
 		RouteMatch match,
 		GlobalInvocationOptions globalOptions,
 		IServiceProvider serviceProvider,
@@ -629,7 +629,7 @@ public sealed partial class CoreReplApp
 		return 1;
 	}
 
-	private async ValueTask<bool> RenderOutputAsync(
+	internal async ValueTask<bool> RenderOutputAsync(
 		object? result,
 		string? requestedFormat,
 		CancellationToken cancellationToken,
@@ -679,7 +679,7 @@ public sealed partial class CoreReplApp
 		return JsonAnsiColorizer.Colorize(payload, _options.Output.ResolvePalette());
 	}
 
-	private async ValueTask<bool> RenderHelpAsync(
+	internal async ValueTask<bool> RenderHelpAsync(
 		GlobalInvocationOptions globalOptions,
 		CancellationToken cancellationToken)
 	{
@@ -802,5 +802,63 @@ public sealed partial class CoreReplApp
 			serviceProvider,
 			_options.Interaction,
 			cancellationToken);
+	}
+
+	private static bool TryFindGlobalCommandOptionCollision(
+		GlobalInvocationOptions globalOptions,
+		HashSet<string> knownOptionNames,
+		out string collidingOption)
+	{
+		foreach (var globalOption in globalOptions.CustomGlobalNamedOptions.Keys)
+		{
+			if (!knownOptionNames.Contains(globalOption))
+			{
+				continue;
+			}
+
+			collidingOption = $"--{globalOption}";
+			return true;
+		}
+
+		collidingOption = string.Empty;
+		return false;
+	}
+
+	private static IReadOnlyDictionary<string, IReadOnlyList<string>> MergeNamedOptions(
+		IReadOnlyDictionary<string, IReadOnlyList<string>> commandNamedOptions,
+		IReadOnlyDictionary<string, IReadOnlyList<string>> globalNamedOptions)
+	{
+		if (globalNamedOptions.Count == 0)
+		{
+			return commandNamedOptions;
+		}
+
+		var merged = new Dictionary<string, IReadOnlyList<string>>(
+			commandNamedOptions,
+			StringComparer.OrdinalIgnoreCase);
+		foreach (var pair in globalNamedOptions)
+		{
+			if (merged.TryGetValue(pair.Key, out var existing))
+			{
+				var appended = existing.Concat(pair.Value).ToArray();
+				merged[pair.Key] = appended;
+				continue;
+			}
+
+			merged[pair.Key] = pair.Value;
+		}
+
+		return merged;
+	}
+
+	private ParsingOptions BuildEffectiveCommandParsingOptions()
+	{
+		var isInteractiveSession = _runtimeState.Value?.IsInteractiveSession == true;
+		return new ParsingOptions
+		{
+			AllowUnknownOptions = _options.Parsing.AllowUnknownOptions,
+			OptionCaseSensitivity = _options.Parsing.OptionCaseSensitivity,
+			AllowResponseFiles = !isInteractiveSession && _options.Parsing.AllowResponseFiles,
+		};
 	}
 }
