--- src/Repl.Core/CoreReplApp.OptionParsing.cs
@@ -1,62 +1,0 @@
-namespace Repl;
-
-public sealed partial class CoreReplApp
-{
-	private static bool TryFindGlobalCommandOptionCollision(
-		GlobalInvocationOptions globalOptions,
-		HashSet<string> knownOptionNames,
-		out string collidingOption)
-	{
-		foreach (var globalOption in globalOptions.CustomGlobalNamedOptions.Keys)
-		{
-			if (!knownOptionNames.Contains(globalOption))
-			{
-				continue;
-			}
-
-			collidingOption = $"--{globalOption}";
-			return true;
-		}
-
-		collidingOption = string.Empty;
-		return false;
-	}
-
-	private static IReadOnlyDictionary<string, IReadOnlyList<string>> MergeNamedOptions(
-		IReadOnlyDictionary<string, IReadOnlyList<string>> commandNamedOptions,
-		IReadOnlyDictionary<string, IReadOnlyList<string>> globalNamedOptions)
-	{
-		if (globalNamedOptions.Count == 0)
-		{
-			return commandNamedOptions;
-		}
-
-		var merged = new Dictionary<string, IReadOnlyList<string>>(
-			commandNamedOptions,
-			StringComparer.OrdinalIgnoreCase);
-		foreach (var pair in globalNamedOptions)
-		{
-			if (merged.TryGetValue(pair.Key, out var existing))
-			{
-				var appended = existing.Concat(pair.Value).ToArray();
-				merged[pair.Key] = appended;
-				continue;
-			}
-
-			merged[pair.Key] = pair.Value;
-		}
-
-		return merged;
-	}
-
-	private ParsingOptions BuildEffectiveCommandParsingOptions()
-	{
-		var isInteractiveSession = _runtimeState.Value?.IsInteractiveSession == true;
-		return new ParsingOptions
-		{
-			AllowUnknownOptions = _options.Parsing.AllowUnknownOptions,
-			OptionCaseSensitivity = _options.Parsing.OptionCaseSensitivity,
-			AllowResponseFiles = !isInteractiveSession && _options.Parsing.AllowResponseFiles,
-		};
-	}
-}
