+++ src/Repl.Core/Routing/ActiveRoutingGraph.cs
@@ -1,0 +1,6 @@
+namespace Repl;
+
+internal readonly record struct ActiveRoutingGraph(
+	RouteDefinition[] Routes,
+	ContextDefinition[] Contexts,
+	ReplRuntimeChannel Channel);
