+++ src/Repl.Core/Routing/PrefixTemplate.cs
@@ -1,0 +1,6 @@
+namespace Repl;
+
+internal readonly record struct PrefixTemplate(
+	RouteTemplate Template,
+	bool IsHidden,
+	IReadOnlyList<string> Aliases);
