+++ src/Repl.Core/Session/AmbientCommandOutcome.cs
@@ -1,0 +1,9 @@
+namespace Repl;
+
+internal enum AmbientCommandOutcome
+{
+	NotHandled,
+	Handled,
+	HandledError,
+	Exit,
+}
